MCUT-Mayhem-Twitch-Extension-Backend/requirements.txt
async-timeout==5.0.1
bidict==0.23.1
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
eventlet==0.39.0
Flask==3.0.3
flask-cors==5.0.1
flask-htmx==0.3.2
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
redis==5.2.1
simple-websocket==1.1.0
SQLAlchemy==2.0.29
typing_extensions==4.12.2
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
Gunicorn